Update modules.scss
On modules.scss, updated line 120 from "a:not(.abut):not(.abut-secondary)" to "a:not(.abut):not(.abut-secondary):not(.calendarEventDateLink)" for calendar list view font color row conflict. Note that there is also a conflict on Mail Reporting of the same nature, but is overwritten with !important white font color around line 1130.